+++ src/main/java/dao/BoardDAO.java
@@ -191,78 +191,6 @@ public class BoardDAO {
 		}
 	}
 
-	// 해당 게시글 찜한 상태인지 확인
-	public boolean isBoardJjim(String seq, String id) throws Exception {
-
-		String sql = "select * from jjim where board_seq = ?  and jjim_id = ?";
-		try (Connection con = this.getConnection(); PreparedStatement pstat = con.prepareStatement(sql)) {
-			pstat.setString(1, seq);
-			pstat.setString(2, id);
-
-			try (ResultSet rs = pstat.executeQuery();) {
-				return rs.next();
-			}
-		}
-	}
-
-	// 해당 게시글 좋아요 상태인지 확인
-	public boolean isBoardGood(String seq, String id) throws Exception {
-
-		String sql = "select * from good where board_seq = ?  and good_id = ?";
-		try (Connection con = this.getConnection(); PreparedStatement pstat = con.prepareStatement(sql)) {
-			pstat.setString(1, seq);
-			pstat.setString(2, id);
-
-			try (ResultSet rs = pstat.executeQuery();) {
-				return rs.next();
-			}
-		}
-	}
-
-	
-	// 애디터추천 All게시글 로그인id가 찜한 정보
-	public List<JjimDTO> selectEditorJjim(String id) throws Exception {
-
-		String sql = "select * from jjim where jjim_id = ? and board_seq like 'e%'";
-		try (Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				ResultSet rs = pstat.executeQuery();) {
-			
-			List<JjimDTO> list = new ArrayList<JjimDTO>();
-			while (rs.next()) {
-				String board_seq = rs.getString("board_seq");
-				String jjim_id = rs.getString("jjim_id");
-				Timestamp jjim_date = rs.getTimestamp("jjimm_date");
-
-				JjimDTO dto = new JjimDTO(board_seq, jjim_id, jjim_date);
-				list.add(dto);
-			}
-			return list;
-		}
-	}
-
-	// 애디터추천 All게시글 로그인id가 좋아요 한 정보
-	public List<GoodDTO> selectEditorGood(String id) throws Exception {
-
-		String sql = "select * from good where good_id = ? and board_seq like 'e%'";
-		try (Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				ResultSet rs = pstat.executeQuery();) {
-			
-			List<GoodDTO> list = new ArrayList<GoodDTO>();
-			while (rs.next()) {
-				String board_seq = rs.getString("board_seq");
-				String good_id = rs.getString("good_id");
-				Timestamp good_date = rs.getTimestamp("good_date");
-
-				GoodDTO dto = new GoodDTO(board_seq, good_id, good_date);
-				list.add(dto);
-			}
-			return list;
-		}
-	}
-	
-	
 	
 	// 조회수 증가
 	public int viewCountUp(String seq) throws Exception {
